+++ src/LDPCDecoders.jl
@@ -15,7 +15,8 @@ export
     BeliefPropagationDecoder,
     BeliefPropagationOSDDecoder,
     BitFlipDecoder,
-    BPOTSDecoder
+    BPOTSDecoder,
+    SmallSetFlipDecoder
 
 include("parity_generator.jl")
 
@@ -24,6 +25,7 @@ include("decoders/belief_propagation.jl")
 include("decoders/belief_propagation_osd.jl")
 include("decoders/iterative_bitflip.jl")
 include("decoders/bpots_decoder.jl")
+include("decoders/small_set_flip.jl")
 
 
 end